fix: expose only answers in question rendering context
